Make trackPathBack depend less on the caller
By inspecting the node that you're looking at instead of the prev, we
allow the caller to pass empty list instead of taking some of the
responsibility of trackPathBack onto itself.

It bothers me a bit that this path seems to have one fewer node in it,
but the behavior is still correct.